+++ packages/expo-router/src/modal/Modal.tsx
@@ -86,6 +86,13 @@ export interface ModalProps extends ViewProps {
    * Works only when `presentation` is set to `formSheet`.
    */
   cornerRadius?: ModalConfig['cornerRadius'];
+  /**
+   * See {@link ScreenProps["unstable_sheetFooter"]}.
+   *
+   * Footer component that can be used alongside formSheet stack presentation style.
+   * Works only when `presentation` is set to `formSheet`.
+   */
+  footer?: ModalConfig['footer'];
 }
 
 /**
@@ -124,8 +131,12 @@ export function Modal(props: ModalProps) {
     presentationStyle,
     transparent,
     detents,
+<<<<<<< HEAD
     detentIndex,
     cornerRadius,
+=======
+    footer,
+>>>>>>> feat/expo-router-modal-footer
     ...viewProps
   } = props;
   const { openModal, updateModal, closeModal, addEventListener } = useModalContext();
@@ -149,7 +160,11 @@ export function Modal(props: ModalProps) {
         parentNavigationProp: navigation,
         detentIndex,
         detents,
+<<<<<<< HEAD
         cornerRadius,
+=======
+        footer,
+>>>>>>> feat/expo-router-modal-footer
       });
       setCurrentModalId(newId);
       return () => {
@@ -164,9 +179,14 @@ export function Modal(props: ModalProps) {
       updateModal(currentModalId, {
         detentIndex,
         component: children,
+        footer,
       });
     }
+<<<<<<< HEAD
   }, [children, detentIndex]);
+=======
+  }, [children, footer]);
+>>>>>>> feat/expo-router-modal-footer
 
   useEffect(() => {
     if (currentModalId) {